requirements.txt
requests>=2.31.0
anthropic>=0.39.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
